renamed time_utils module to datetime_utils
