+++ src/lib/metrics_extractor/glue_workflows_metrics_extractor.py
@@ -2,7 +2,7 @@ from datetime import datetime
 from lib.metrics_extractor.base_metrics_extractor import BaseMetricsExtractor
 
 from lib.aws.glue_manager import GlueManager, WorkflowRun
-from lib.core import time_utils
+from lib.core import datetime_utils
 
 
 class GlueWorkflowsMetricExtractor(BaseMetricsExtractor):
@@ -72,7 +72,7 @@ class GlueWorkflowsMetricExtractor(BaseMetricsExtractor):
                     for metric_name, metric_value, metric_type in metric_values
                 ]
 
-                record_time = time_utils.datetime_to_epoch_milliseconds(
+                record_time = datetime_utils.datetime_to_epoch_milliseconds(
                     workflow_run.StartedOn
                 )
